WIP on merging coalescence functionality
