+++ src/Cloudy.jl
@@ -1,10 +1,10 @@
 module Cloudy
 
-include(joinpath("Sources","EquationTypes.jl"))
 include(joinpath("Kernels","KernelTensors.jl"))
 include(joinpath("Kernels","KernelFunctions.jl"))
 include(joinpath("ParticleDistributions","ParticleDistributions.jl"))
-include(joinpath("Sources","Sources.jl"))
-include(joinpath("Sources","MultiParticleSources.jl"))
+include(joinpath("Sources","EquationTypes.jl"))
+include(joinpath("Sources","Coalescence.jl"))
+include(joinpath("Sources","Sedimentation.jl"))
 
 end
