+++ src/Sources/EquationTypes.jl
@@ -6,12 +6,18 @@ module EquationTypes
 
 export AbstractStyle
 export CoalescenceStyle
-export OneModeCoalStyle
-export TwoModesCoalStyle
+export AnalyticalCoalStyle
+export NumericalCoalStyle
+export HybridCoalStyle # TODO
+export OneModeCoalStyle # TODO: combine with TwoMode to become "Analytical"
+export TwoModesCoalStyle # TODO: combine with OneMode to become "Analytical"
 
 abstract type AbstractStyle end
 abstract type CoalescenceStyle <: AbstractStyle end
-struct OneModeCoalStyle <: CoalescenceStyle end
-struct TwoModesCoalStyle <: CoalescenceStyle end
+struct AnalyticalCoalStyle <: CoalescenceStyle end
+struct NumericalCoalStyle <: CoalescenceStyle end
+struct HybridCoalStyle <: CoalescenceStyle end
+struct OneModeCoalStyle <: CoalescenceStyle end # TODO: remove
+struct TwoModesCoalStyle <: CoalescenceStyle end # TODO: remove
 
 end #module EquationTypes.jl
